Move sponsored podcasts loading to view model
